Tried Raphael, but it was too slow. Canvas is next
The Raphael code was taking 1.3 seconds to render the same thing that
the canvas code would render in 5ms. It was way too slow for our needs.
Now just need to figure out how to get the canvas to scroll and zoom.
And get the text in the right place.