Got the text positioning working in Firefox
This is a major step forward. The problem with the positioning was that
the object was being placed at the correct position before being
transformed and then it was being moved after the transform. This was
resulting in a lot of weird offsets that I couldn't figure out and still
don't really understand. The answer was to draw all the letters at 0,0
and then use the transform to position them in the correct place after
scaling.