Created logo object and used multiple canvases
There is now an HMMLogo object that is used to render the graphics. It
also decides if text or rectangles will be used to represent the items in
each column. Finally the zoom feature is working but it re-renders in a
way that will leave you disorientated. Need to figure out how to
maintain a center point.