Added AA colors and fixed EasyScroller init.
The EasyScroller object was getting initialized twice instead of just
once, because my code was calling the init methods on page load and so
was the code at the end of the EasyScroller.js library. I removed the
attributes from the DOM element that triggered the code at the end of
EasyScroller.js and added my own trigger for my custom method.