Added scaled_max option.
When creating an hmm logo the scaled_max option allows you to scale the
height to the theoretical max or the observed max for that logo. This
allows one to maximise the height of the letters in the logo or render
them all according to the same scale to allow direct visual comparisons
between logos.