Adds max height toggling to the logos
There are now two different heights being passed in the json string.
One is the maximum theoretical height and the other is the maximum
observed height. This update checks if those values have been passed
in at object creation and if they have, then a scale toggle switch is
added to the output to allow toggling between them.