Caching the canvas supported check
It appears that checking if the canvas is supported all the time is a
bad thing to do. It was attaching a canvas object and checking the
return value many times over. Since the canvas support is unlikely to
change after a browser has loaded a page, I have cached the result after
the first run and it seems to have greatly improved the speed of the
code. Next time I should look at the profiler output before waisting a
day trying to fix other parts of the code that I think are slow.