Fixed buggy scale toggling
Sometimes, when toggling the scale used to determine the height of the
logo, I noticed that axis labels were changing, but the actual canvas
was not. This is because the canvas was not being rendered again due to
an optimisation I added to prevent canvas rendering if the position on
the screen hadn't changed. To fix it I just shifted over one column and
then back again to force a render.