Removed ids as they wont allow multiple logos
Having ids as hooks is fine if you only intend to have one logo on a
page. Since I am trying to make this generic, then I need to make sure
that people can use it multiple times in a page by using classes and
more specific selectors for events.