JSON data now an attribute of target div.
To try and make things easier to embed into another site, I have removed
the need to put the data inside a script tag as a stand alone variable.
It is now attached to the target div as a data-logo attribute. This
means that the end user only has to insert a div with the correct
attributes into their page and call the .hmm_logo function to get a
working logo.