Added 'column_info' option.
The column_info option allows one to pass in a selector for an element
on the page that will have its' contents replaced with a table of all
the residues in a particular column, when that column is clicked. This
element can be styled anyway one likes and placed anywhere on the page.
If not passed, then the onclick event will do nothing.