Fix reversing columns in the logo on click.
I had to clone the column data that was being shown in the column info
div as reversing it was messing up the logo when zoom levels change. The
column that had been reversed for the info box was also being reversed
in the graphic, so it looked out of place. It was also flip-flopping the
order from ascending to descending in a non-intuitive manner.