Column info table split into 4 columns
The initial iteration of this table was to show all the data in one long
column. This was not ideal for including on a web page as it took up a
lot of vertical space and left a lot of white space on the page. So, It
has been split into columns of 5 values or less, so that it takes up
more horizontal space, but less vertical space.