Removed vertical blue line after delete.
The delete probability section was copied from the insert probability
section wholesale. This included a vertical line after the columns where
there was a high probability of a delete occurring. The delete would be a
removal of that column, so it is best to just remove the line, since the
vertical line was intended to indicate that an insertion would happen after
the column, it doesn't map to a delete correctly.