Reverting canvas copying of text to generate letters.
Although the copying method was faster and produced letters in the
correct location that didn't overlap, it looked a lot rougher. It was
harder to tell what the smaller letters were supposed to be and the
larger letters were not very smooth. This reverts back to using the
nicely anti aliased letters from before, with some improvements for
rendering individual letters.